Recast populated space to integers

+++ src/respace/result.py
@@ -593,7 +593,7 @@ class ResultSet:
 
     @property
     def populated_space(self) -> xr.Dataset:
-        return self.param_space.where(self.populated_mask, drop=True)
+        return self.param_space.where(self.populated_mask, drop=True).astype("int")
 
     def get_subspace_res(
         self,
